Three changes to dealing with @addcommand'd stuff in process_cmdent:
When matching addcommanded command strings, the string to match against
was being unnecessarily repeatedly reconstructed. This has been fixed.

New conf parameter, addcommands_match_blindly, controls whether or not
a 'Huh?' is produced when a command is @addcommand'd but the syntax
typed by the player does not match the wildcard $commands on any of
the relevant attributes. In the past, this failed silently (as it does
now with the default of 'yes').

New conf parameter, addcommands_obey_stop, controls whether or not
further matching attempts are made when an @addcommand'd command is
matched and executed, if the object it's on is set STOP. In the past,
it didn't (as it does now with the default of 'no'); setting it to
'yes' is consistent with the normal behavior of objects and is an
efficiency win in cases where there's a lot of attributes associated
with a single command (like say, 'foo/x1', 'foo/x2', 'foo/x3', etc.)