Interim step towards module database interactions.
dbconvert now is able to load modules. it reads the config file ('module' is
the only directive it understands, it ignores all others). It performs no
module initialization. This is intended to allow dbconvert to call module
routines to flatfile/load module data to and from GDBM. This required
changing the dbconvert option convention; dbconvert now uses getopt() and
uses -c for the config file name. Backup and Restore scripts have been
changed to reflect this.

The way I have planned this API is to allow modules to register with the
MUSH for a database type (an integer; see udb_defs.h). This will be
dynamically allocated and the module to type mapping will be stored in GDBM.
The module will then use this type to store items in the cache/GDBM without
conflicting with regular MUSH entries, since type is part of the key. The
API will also have calls for flatfiling; basically the MUSH will feed the
module a line from the flatfile, and the module will have to know what to do
with it. The MUSH will ignore lines from the flatfile which are for modules
it doesn't know about. For the inverse, the module will need to keep track
of what keys it has stored in GDBM, and convert them to ASCII lines which it
then feeds to the MUSH to insert into the flatfile.

I hope to have this finished this weekend. ;)