Creation timestamps and the creation() function.
WARNING: This requires a change in database format (to add the line that
contains the create date). Shut down and run a backup BEFORE you build
the new source, build, restore from backup using the NEW code, then
start the game. DO NOT APPLY THIS UPDATE LIVE with just a @restart!